#71 🪲 BUG: 'Reduce max. (dis-)charge power' setting not preserved
This was due to incomplete list of 'True values' in the __check_and_convert_value method.
Also fixed:
- Different FlexMeasures URL can now be used and remains valid over restarts.
- VAT and Markup can be used.
For all optional values: the content of the input-fields is saved if toggle is switched off.

***Please note***
Somehow a lot of changes from bug #84 'Getting EPEX prices on daily schedule fails' had to be re-commited (wrong base branch?)