#115 FR: Make getting prices more robust
 Now, if the getting the prices fails or is not up to date yet, we check every hour until the following noon.
There is one check at 18:30 to see if prices are up to date, if not the user is notified.
The notification is removed if prices are up to date again.
The code is consolidated where possible.
The test version of get_fm_data.py is added to a tests folder.