Refactor `DataMonitor` module to receive all changes to eventbus.
For this the `charger_state_change` event is extended with the old state.

For clarity the `update_charger_connection_state` event is renamed to `charger_communication_state_change` as it was easy to confuse with `is_car_connected`.
Removed some code that has become obsolete due to the refactoring.
Added `add_keep_alive` parameter to `__update_ha_entity` method in preparation for more situations that need this functionality.